dev_024: include loan charge details in repayment schedule api
