+++ fineract-provider/src/main/java/org/apache/fineract/portfolio/loanaccount/service/LoanChargeAssembler.java
@@ -26,16 +26,21 @@ import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.HashSet;
+import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Locale;
 import java.util.Set;
+import java.util.stream.Collectors;
+
 import lombok.RequiredArgsConstructor;
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.fineract.infrastructure.core.api.JsonCommand;
+import org.apache.fineract.infrastructure.core.data.EnumOptionData;
 import org.apache.fineract.infrastructure.core.domain.ExternalId;
 import org.apache.fineract.infrastructure.core.serialization.FromJsonHelper;
 import org.apache.fineract.infrastructure.core.service.DateUtils;
 import org.apache.fineract.infrastructure.core.service.ExternalIdFactory;
+import org.apache.fineract.organisation.monetary.data.CurrencyData;
 import org.apache.fineract.portfolio.charge.domain.Charge;
 import org.apache.fineract.portfolio.charge.domain.ChargeCalculationType;
 import org.apache.fineract.portfolio.charge.domain.ChargePaymentMode;
@@ -43,11 +48,15 @@ import org.apache.fineract.portfolio.charge.domain.ChargeRepositoryWrapper;
 import org.apache.fineract.portfolio.charge.domain.ChargeTimeType;
 import org.apache.fineract.portfolio.charge.exception.LoanChargeNotFoundException;
 import org.apache.fineract.portfolio.charge.exception.LoanChargeWithoutMandatoryFieldException;
+import org.apache.fineract.portfolio.charge.service.ChargeEnumerations;
 import org.apache.fineract.portfolio.loanaccount.api.LoanApiConstants;
+import org.apache.fineract.portfolio.loanaccount.data.LoanChargeData;
+import org.apache.fineract.portfolio.loanaccount.data.LoanInstallmentChargeData;
 import org.apache.fineract.portfolio.loanaccount.domain.Loan;
 import org.apache.fineract.portfolio.loanaccount.domain.LoanCharge;
 import org.apache.fineract.portfolio.loanaccount.domain.LoanChargeRepository;
 import org.apache.fineract.portfolio.loanaccount.domain.LoanDisbursementDetails;
+import org.apache.fineract.portfolio.loanaccount.domain.LoanInstallmentCharge;
 import org.apache.fineract.portfolio.loanaccount.domain.LoanTrancheDisbursementCharge;
 import org.apache.fineract.portfolio.loanproduct.domain.LoanProduct;
 import org.apache.fineract.portfolio.loanproduct.domain.LoanProductRepository;
@@ -328,4 +337,38 @@ public class LoanChargeAssembler {
         return loanChargeService.create(null, chargeDefinition, loanPrincipal, amount, chargeTime, chargeCalculation, dueDate,
                 chargePaymentMode, numberOfRepayments, BigDecimal.ZERO, externalId);
     }
+
+    public Set<LoanChargeData> convertLoanChargesToData(Set<LoanCharge> loanCharges) {
+        if (loanCharges == null || loanCharges.isEmpty()) {
+            return new LinkedHashSet<>();
+        }
+        return loanCharges.stream().map(this::mapToLoanChargeData).collect(Collectors.toCollection(LinkedHashSet::new));
+    }
+
+    private LoanChargeData mapToLoanChargeData(LoanCharge loanCharge) {
+        Charge chargeDefinition = loanCharge.getCharge();
+        if (chargeDefinition == null) {
+            return loanCharge.toData();
+        }
+
+        CurrencyData currency = chargeDefinition.toData().getCurrency();
+        EnumOptionData chargeTimeType = ChargeEnumerations.chargeTimeType(chargeDefinition.getChargeTimeType());
+        EnumOptionData chargeCalculationType = ChargeEnumerations.chargeCalculationType(chargeDefinition.getChargeCalculation());
+        EnumOptionData chargePaymentMode = ChargeEnumerations.chargePaymentMode(chargeDefinition.getChargePaymentMode());
+        List<LoanInstallmentChargeData> installmentChargeData = loanCharge.installmentCharges().stream().map(LoanInstallmentCharge::toData)
+                .toList();
+
+        return LoanChargeData.builder().id(loanCharge.getId()).chargeId(chargeDefinition.getId()).name(chargeDefinition.getName())
+                .currency(currency).amount(loanCharge.getAmount()).amountPaid(loanCharge.getAmountPaid())
+                .amountWaived(loanCharge.getAmountWaived()).amountWrittenOff(loanCharge.getAmountWrittenOff())
+                .amountOutstanding(loanCharge.getAmountOutstanding()).chargeTimeType(chargeTimeType)
+                .submittedOnDate(loanCharge.getSubmittedOnDate()).dueDate(loanCharge.getDueLocalDate())
+                .chargeCalculationType(chargeCalculationType).percentage(loanCharge.getPercentage())
+                .amountPercentageAppliedTo(loanCharge.getAmountPercentageAppliedTo()).amountOrPercentage(loanCharge.getAmountOrPercentage())
+                .penalty(chargeDefinition.isPenalty()).chargePaymentMode(chargePaymentMode).paid(loanCharge.isPaid())
+                .waived(loanCharge.isWaived()).loanId(loanCharge.getLoan() != null ? loanCharge.getLoan().getId() : null)
+                .minCap(chargeDefinition.getMinCap()).maxCap(chargeDefinition.getMaxCap()).installmentChargeData(installmentChargeData)
+                .externalId(loanCharge.getExternalId()).build();
+    }
+
 }
